app/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
fastapi==0.63.0
h11==0.12.0
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
pyjwt==2.0.0
pymongo==3.12.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
uvicorn==0.13.3
google-cloud-pubsub==2.7.1